chore(Portal): Move /src/core into /src/shared
