+++ packages/dnb-design-system-portal/gatsby-config.js
@@ -6,7 +6,7 @@
 const remarkGfm = require('remark-gfm')
 const getCurrentBranchName = require('current-git-branch')
 const currentBranch = getCurrentBranchName()
-const { shouldUsePrebuild } = require('./src/core/BuildTools.cjs')
+const { shouldUsePrebuild } = require('./src/shared/core/BuildTools.cjs')
 
 const pathPrefix = '/'
