+++ packages/dnb-design-system-portal/gatsby-ssr.tsx
@@ -3,8 +3,11 @@
  *
  */
 
-import { rootElement, pageElement } from './src/core/PortalProviders'
-import { renderBody } from './src/core/PortalHead'
+import {
+  rootElement,
+  pageElement,
+} from './src/shared/core/PortalProviders'
+import { renderBody } from './src/shared/core/PortalHead'
 
 export const wrapRootElement = rootElement('ssr')
 export const wrapPageElement = pageElement()
