+++ packages/dnb-design-system-portal/src/docs/quickguide-designer/colors-table.mdx
@@ -1,6 +1,6 @@
 import Table from 'dnb-design-system-portal/src/shared/tags/Table'
 import { Space } from '@dnb/eufemia/src'
-import ChangeStyleTheme from '../../core/ChangeStyleTheme'
+import ChangeStyleTheme from 'dnb-design-system-portal/src/shared/core/ChangeStyleTheme'
 
 Change the brand in order to see the related tables:
