+++ packages/dnb-design-system-portal/src/docs/uilib/components/date-picker/demos.mdx
@@ -18,7 +18,7 @@ import {
   DatePickerScreenshotTestSizes,
   DatePickerScreenshotTestDisabled,
 } from 'Docs/uilib/components/date-picker/Examples'
-import ChangeLocale from 'dnb-design-system-portal/src/core/ChangeLocale'
+import ChangeLocale from 'dnb-design-system-portal/src/shared/core/ChangeLocale'
 
 ## Demos
