+++ packages/dnb-design-system-portal/src/docs/uilib/components/input-masked/demos.mdx
@@ -11,7 +11,7 @@ import {
   InputMaskedExamplePrefix,
   InputMaskedExamplePhone,
 } from 'Docs/uilib/components/input-masked/Examples'
-import ChangeLocale from 'dnb-design-system-portal/src/core/ChangeLocale'
+import ChangeLocale from 'dnb-design-system-portal/src/shared/core/ChangeLocale'
 
 ## Demos
