+++ packages/dnb-design-system-portal/src/docs/uilib/components/number-format/demos.mdx
@@ -17,7 +17,7 @@ import {
   NumberSpacing,
   NumberProvider,
 } from 'Docs/uilib/components/number-format/Examples'
-import ChangeLocale from 'dnb-design-system-portal/src/core/ChangeLocale'
+import ChangeLocale from 'dnb-design-system-portal/src/shared/core/ChangeLocale'
 
 ## Demos
