+++ packages/dnb-design-system-portal/src/docs/uilib/components/skeleton/demos.mdx
@@ -11,7 +11,7 @@ import {
   SkeletonVisualTests,
 } from 'Docs/uilib/components/skeleton/Examples'
 
-import PortalSkeleton from 'dnb-design-system-portal/src/core/PortalSkeleton'
+import PortalSkeleton from 'dnb-design-system-portal/src/shared/core/PortalSkeleton'
 
 ## Demos
