+++ packages/dnb-design-system-portal/src/docs/uilib/extensions/payment-card/demos.mdx
@@ -8,7 +8,7 @@ import PaymentCard, {
   ProductType,
   CardType,
 } from '@dnb/eufemia/src/extensions/payment-card'
-import ChangeLocale from 'dnb-design-system-portal/src/core/ChangeLocale'
+import ChangeLocale from 'dnb-design-system-portal/src/shared/core/ChangeLocale'
 import {
   PaymentCardBasicExample,
   PaymentCardCustomExample,
