+++ packages/dnb-design-system-portal/src/docs/uilib/usage/customisation/theming/theme/demos.mdx
@@ -6,7 +6,7 @@ import {
   ThemeBasis,
   ThemeMapping,
 } from 'Docs/uilib/usage/customisation/theming/theme/Examples'
-import ChangeStyleTheme from '../../../../../../core/ChangeStyleTheme'
+import ChangeStyleTheme from 'dnb-design-system-portal/src/shared/core/ChangeStyleTheme'
 
 ## Demos
