+++ packages/dnb-design-system-portal/src/docs_dummy/dummy.mdx
@@ -19,7 +19,7 @@ redirect_from:
   - /other-dummy
 ---
 
-import ChangeStyleTheme from '../core/ChangeStyleTheme'
+import ChangeStyleTheme from 'dnb-design-system-portal/src/shared/core/ChangeStyleTheme'
 
 # What is this?
