+++ packages/dnb-design-system-portal/src/shared/core/BuildTools.cjs
