+++ packages/dnb-design-system-portal/src/shared/core/PortalSkeleton.tsx
